AR-611: make group-columns.jsx
